+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bug fixes
  * Fixed CSRF Protection in PHP build
+ * UI/UX updated in signature boxes in service forms
 
 ### Features
  * Added About page link to the footer